# Conflicts:
#	frontend/src/App.js
#	frontend/src/Header.js
#	src/main/java/by/itech/lab/supplier/controller/UserController.java
#	src/main/java/by/itech/lab/supplier/dto/UserDto.java

+++ src/main/java/by/itech/lab/supplier/service/impl/ItemServiceImpl.java
@@ -5,6 +5,7 @@ import by.itech.lab.supplier.dto.CategoryDto;
 import by.itech.lab.supplier.dto.ItemDto;
 import by.itech.lab.supplier.dto.mapper.ItemMapper;
 import by.itech.lab.supplier.exception.ResourceNotFoundException;
+import by.itech.lab.supplier.repository.ApplicationItemRepository;
 import by.itech.lab.supplier.repository.ItemRepository;
 import by.itech.lab.supplier.service.CategoryService;
 import by.itech.lab.supplier.service.ItemService;
@@ -15,7 +16,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
@@ -25,17 +28,16 @@ public class ItemServiceImpl implements ItemService {
 
     private final ItemMapper itemMapper;
 
-    private final CategoryService categoryService;
+    private final ApplicationItemRepository applicationItemRepository;
 
     @Override
     public Page<ItemDto> findByLabel(final String label, final Pageable pageable) {
         return itemRepository.findByLabel(label, pageable)
-          .map(itemMapper::map);
+                .map(itemMapper::map);
     }
 
-    public Page<ItemDto> findAllByCategory(final String categoryName, final Pageable pageable) {
-        CategoryDto found = categoryService.findByCategory(categoryName);
-        return itemRepository.findAllByCategory(found.getId(), pageable)
+    public Page<ItemDto> findAllByCategory(final CategoryDto categoryDto, final Pageable pageable) {
+        return itemRepository.findAllByCategory(categoryDto.getId(), pageable)
           .map(itemMapper::map);
     }
 
@@ -43,34 +45,44 @@ public class ItemServiceImpl implements ItemService {
     @Transactional
     public ItemDto save(final ItemDto dto) {
         Item item = Optional.ofNullable(dto.getId())
-          .map(itemToSave -> {
-              final Item existing = itemRepository
-                .findById(dto.getId())
-                .orElseThrow();
-              itemMapper.map(dto, existing);
-              return existing;
-          })
-          .orElseGet(() -> itemMapper.map(dto));
+                .map(itemToSave -> {
+                    final Item existing = itemRepository
+                            .findById(dto.getId())
+                            .orElseThrow();
+                    itemMapper.map(dto, existing);
+                    return existing;
+                })
+                .orElseGet(() -> itemMapper.map(dto));
 
         final Item saved = itemRepository.save(item);
         return itemMapper.map(saved);
     }
 
     @Override
-    public Page<ItemDto> findAll(Pageable pageable) {
+    public Page<ItemDto> findAll(final Pageable pageable) {
         return itemRepository.findAll(pageable)
-          .map(itemMapper::map);
+                .map(itemMapper::map);
     }
 
     public ItemDto findById(final Long id) {
         return itemRepository.findById(id).map(itemMapper::map)
-          .orElseThrow(() -> new ResourceNotFoundException("Item with id=" + id + " doesn't exist"));
+                .orElseThrow(() -> new ResourceNotFoundException("Item with id=" + id + " doesn't exist"));
     }
 
     @Override
     @Transactional
-    public void delete(Long id) {
-        itemRepository.deleteById(id, LocalDate.now());
+    public void delete(final Long id) {
+        if (applicationItemRepository.getNumberOfItemUsages(id) == 0) {
+            itemRepository.deleteById(id, LocalDate.now());
+        } else {
+            throw new ResourceNotFoundException("Item can not be deleted, because it is already used in " +
+                    "application");
+        }
+    }
+
+    @Override
+    public List<ItemDto> findByUpc(String upc) {
+        return itemRepository.findByUpcStartsWith(upc).stream().map(itemMapper::map).collect(Collectors.toList());
     }
 
 }
